fix(settings): track language choice via a dedicated defaults key
AppleLanguages lives in the global domain, so reading it back can't distinguish a user-set override from the inherited system language (the picker would show a language as selected even when set to follow the system). Record the choice under DockyPreferredLanguage as the source of truth, while still writing AppleLanguages for macOS to honor at launch. Split DisplayLanguage/LanguageController into their own file so the logic compiles into the test target without the SwiftUI view.
